fork of the stable-5 branch for game_intllib
This is a fork of the 5.0.1 version i.e. stable-5 branch for the limited purpose of the game_intllib project. This project is aiming at LUA-code translation in compliance to the GNU gettext toolchain and intentionally has the limited scope of a 'one-lamguage' server solution. This means the server 'language=' settings chooses one main language and both the translation of S(string") tags in the LUA-code is server driven.